requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.1.7
django-admin-interface==0.20.0
django-autoslug==1.9.8
django-ckeditor==6.4.2
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-etc==1.4.0
django-extensions==3.1.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-hitcount==1.3.5
django-js-asset==2.0.0
django-social-share==2.3.0
django-summernote==0.8.20.0
django-taggit==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
urllib3==1.26.14
webencodings==0.5.1
